relocate serialization.py to avoid circular import
