+++ tensorrt_llm/executor/proxy.py
@@ -9,7 +9,7 @@ import torch
 import zmq
 import zmq.asyncio
 
-import tensorrt_llm.executor.serialization as serialization
+import tensorrt_llm.serialization as serialization
 from tensorrt_llm.logger import logger
 
 from .._utils import mpi_rank
