+++ tensorrt_llm/serialization.py
@@ -97,7 +97,9 @@ BASE_ZMQ_CLASSES = {
 }
 
 BASE_PARALLEL_CLASSES = {
-    "tensorrt_llm.auto_parallel.parallelization": ["ParallelConfig"]
+    "tensorrt_llm.auto_parallel.parallelization": ["ParallelConfig"],
+    "tensorrt_llm.auto_parallel.config": ["AutoParallelConfig", "CostModel"],
+    "tensorrt_llm.auto_parallel.simplifier": ["GraphConfig", "StageType"]
 }
